Added Search Functions
Exchanged the Character List for a Search for Characters window, so that
you may list all characters or search for specific ones by name or ID.

Also I'm sorry, but I tweaked the name-searching function in the dnd
module slightly. This was done so it would turn everything to lowercase
when it was compared so that you didn't have to be 100% correct when
searching so that I could write "aust liagala" and still get my
character "Aust Liagala" from the database.