Refined Searching and character creation
Fixed so you can now see all characters of the same name.
Made new function that will capitalize the first letter of a string.
Especially handy when dealing with entering new classes, as the stock
classes are capitalized.
also fixed it so you can now submit the form without it refreshing the
page.